+++ tarantool/src/lib.rs
@@ -176,7 +176,6 @@ pub mod space;
 pub mod sql;
 #[cfg(feature = "tarantool_test")]
 pub mod test;
-pub mod test_utils;
 pub mod transaction;
 pub mod trigger;
 pub mod tuple;