+++ tests/src/common.rs
@@ -1,7 +1,5 @@
 use serde::{Deserialize, Serialize};
 
-pub use tarantool::test_utils::{check_yield, fiber_csw, YieldResult};
-
 use tarantool::{
     tlua::{self, AsLua, LuaState},
     tuple::Encode,